requirements.txt
pymongo>=4.0.0
python-dotenv>=0.19.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.0
